Updates 'squad' state name to 'faceoffs'
