+++ src/components/pages/VotingPage/VotingPageContainer.js
@@ -15,7 +15,6 @@ function VotingPageContainer({ LoadingComponent, ...props }) {
   const [memoAuthService] = useMemo(() => [authService], []);
   const [faceoff, setFaceoff] = useState();
 
-
   useEffect(() => {
     let isSubscribed = true;
 
@@ -36,20 +35,24 @@ function VotingPageContainer({ LoadingComponent, ...props }) {
   }, [memoAuthService]);
 
   useEffect(() => {
-    getGameVotes(authState, props.squad[0].SquadID, props.child.memberId).then(res => {
+    getGameVotes(
+      authState,
+      props.faceoffs[0].SquadID,
+      props.child.memberId
+    ).then(res => {
       if (res.length === 0) {
-        setFaceoff(props.squad[3]);
+        setFaceoff(props.faceoffs[3]);
       } else if (res.length === 1) {
-        setFaceoff(props.squad[2]);
+        setFaceoff(props.faceoffs[2]);
       } else if (res.length === 2) {
-        setFaceoff(props.squad[1]);
+        setFaceoff(props.faceoffs[1]);
       } else if (res.length === 3) {
-        setFaceoff(props.squad[0]);
+        setFaceoff(props.faceoffs[0]);
       } else {
         push('/child/dashboard');
       }
     });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
@@ -72,7 +75,7 @@ function VotingPageContainer({ LoadingComponent, ...props }) {
 export default connect(
   state => ({
     child: state.child,
-    squad: state.squad,
+    faceoffs: state.faceoff,
   }),
   {}
 )(VotingPageContainer);
