+++ src/state/actions/faceoffsActions.js
@@ -1,4 +1,4 @@
 export const SET_SQUAD_FACEOFFS = 'SET_SQUAD_FACEOFFS';
 export const setSquadFaceoffs = squadInformation => dispatch => {
-    dispatch({ type: SET_SQUAD_FACEOFFS, payload: squadInformation });
+  dispatch({ type: SET_SQUAD_FACEOFFS, payload: squadInformation });
 };
