+++ src/state/actions/index.js
@@ -3,6 +3,6 @@ import * as child from './childActions';
 import * as parent from './parentActions';
 import * as tasks from './taskActions';
 import * as team from './teamActions';
-import * as squad from './squadActions';
+import * as faceoffs from './faceoffsActions';
 
-export { global, child, parent, tasks, team, squad }; 
+export { global, child, parent, tasks, team, faceoffs };
